시간지나면 스텟깎임

+++ Zelter23Day/01_WinMain/PlayerHPUI.h
@@ -10,7 +10,7 @@ class PlayerHPUI : public UI
 	Image* mFullGaugeImage;
 	Image* mHeartImage;
 
-	int mHP;
+	float mHP;
 
 	Animation* mHeartAnimation;
 public:
